pyproject.toml
[tool.poetry]
name = "news2img"
version = "0.1.0"
description = "Generating images of recent news according to user's mood."
authors = [
    "Mingyu Gao <231250154@smail.nju.edu.cn>",
    "Yuping Liu <221220104@smail.nju.edu.cn>",
    "Hongzhi Mao <hongzhimao@smail.nju.edu.cn>",
    "Fanjun Zeng <221220095@smail.nju.edu.cn>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
feedparser = "^6.0.11"
configargparse = "^1.7"
platformdirs = "^4.3.6"
loguru = "^0.7.3"
opencv-python = "^4.10.0.84"
numpy = "^1.26"
transformers = "^4.47.0"
pillow = "^11.0.0"
diffusers = "^0.31.0"
torch = "^2.5.1"
tiktoken = "^0.8.0"
openai = "^1.57.0"
httpx = "^0.28.1"

[[tool.poetry.source]]
name = "nju-mirror"
url = "https://mirror.nju.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.2"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"

[tool.poe]
poetry_command = ""
envfile = ".env"

[tool.poe.tasks]
exec = "python3 -m news2img"
test = "pytest"
lint = "ruff check news2img"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
